requirements.txt
ipykernel
numpy 
pandas
openpyxl
catboost
xgboost
scikit-learn
joblib
dvc
python-dotenv
python-box


# EDA Requirements (Can be ignored by dockerfile, because they are not necessary to run the CICD Pipeline)
ydata-profiling
matplotlib
seaborn


